WIP: add data migration to change regex
WARNING: this migration is broken, run the test of the migration to
figure out why.

The plan is to add '\b' or similar before and after the regex, in order
to limit the boundaries of a word.  So strings like 'i#1', '[i#1]', '
i#1', etc. match but strings like 'gh#cli/cli#1' don't match with the
regular expression "i#(\d+)".

TODO: we need to confirm that this regex works fine with Perl in the backend.