Use new receiver role names for newly created event subscriptions
In order to migrate the receiver role names from `watcher`
to `project_watcher` we have to first use the new receiver role
names for all newly created subscriptions. For now we still
support the old naming by keeping the old receiver role names
for the events, so we can still match them against each other.
So we going to support both until we migrated all existing
event subscriptions in a further step.