Validate supported filter for event on workflow execution
Right now we have an intermix we the validations for a SCM
workflow. When we intialize the `Workflow` class we run
the `WorkflowFiltersValidator`. This validator makes sure that
all filters in the `Workflow` are valid (so basically that every
filter definition in the `workflow.yml` file are correct).

The problem right now is, we cannot check in this validator that
the current webhook event is supported by the filters (in this
case the branches filter for the `tag_push`) since the class is
always intialized no matter if it is meant for the currently incoming
webhook.

If a certain webhook event is meant for a the workflow is checked
when we execute the workflow (`Workflow.call`). Here we check
if the event matches the event filter etc.
Therefore we also have to check here if the current event supports
the filters or not.

An issue occurs currently when there a multiple workflow definitions
in the same `workflow.yml` file. For example:

```
rebuild:
  steps:
    - branch_package:
        source_project: home:krauselukas
        source_package: hello_world
        target_project: home:krauselukas:test
  filters:
    event: push
    branches:
      only:
        - main
release:
  steps:
    - trigger_services:
        project: home:krauselukas
        package: hello_world
  filters:
    event: tag_push
```

The first workflow is for the `push` event, which supports
the `branches` filter. The second for `tag_push` doesn't.
This will fail, when the event is a `tag_push` since the
`WorkflowFiltersValidator` is called when intializing the workflow
and checks the current webhook event against all fitlers (which in
this example also includes a branches fitler).

With the current architecture we have to validate for this
when the workflow is executed.